fix(long_overflow): AND-0000 Eth Balance Overflow (#3234)

+++ wallet/src/main/java/info/blockchain/wallet/ethereum/util/EthUtils.kt
@@ -4,7 +4,7 @@ import org.spongycastle.util.encoders.Hex
 
 object EthUtils {
 
-    fun convertHexToBigInteger(hex: String) = hex.removePrefix(PREFIX).toLong(RADIX).toBigInteger()
+    fun convertHexToBigInteger(hex: String) = hex.removePrefix(PREFIX).toBigInteger(RADIX)
 
     fun decorateAndEncode(signedBytes: ByteArray) = PREFIX + String(Hex.encode(signedBytes))
